+++ packages/adminui-vue/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - `UToolbar`: the `dropdown-prepend` slot to add some items at the start of dropdown
 
 ### Changed
+ - `UTable/UTableEntity/UCardView`: improved mobile UX. Click\tap on active cell\card now open edit form (like double click on row) 
 
 ### Deprecated